+++ python/pragzip/pragzip.pyx
@@ -6,6 +6,7 @@ from libc.stdlib cimport malloc, free
 from libc.stdio cimport SEEK_SET
 from libcpp.string cimport string
 from libcpp.map cimport map
+from libcpp.vector cimport vector
 from libcpp cimport bool
 from cpython.buffer cimport PyObject_GetBuffer, PyBuffer_Release, PyBUF_ANY_CONTIGUOUS, PyBUF_SIMPLE
 from cpython.ref cimport PyObject
@@ -18,8 +19,10 @@ import sys
 ctypedef (unsigned long long int) size_t
 ctypedef (long long int) lli
 
+cdef extern from "tools/pragzip.cpp":
+    int pragzipCLI(int, char**) except +
 
-cdef extern from "ParallelGzipReader.hpp":
+cdef extern from "pragzip/ParallelGzipReader.hpp":
     cppclass ParallelGzipReader:
         ParallelGzipReader(string, size_t) except +
         ParallelGzipReader(int, size_t) except +
@@ -235,6 +238,24 @@ def open(filename, parallelization = 0):
               with suitable read, seekable, seek, and tell methods.
     """
     return PragzipFile(filename, parallelization)
+
+
+def cli():
+    args = sys.argv
+    cdef char** cargs = <char**> malloc(len(args) * sizeof(char*))
+    cdef vector[Py_buffer] buffers
+    buffers.resize(len(args))
+
+    try:
+        for i, arg in enumerate(args):
+            PyObject_GetBuffer(arg.encode(), &buffers[i], PyBUF_SIMPLE | PyBUF_ANY_CONTIGUOUS)
+            cargs[i] = <char*>buffers[i].buf
+
+        return pragzipCLI(len(args), cargs)
+    finally:
+        free(cargs)
+        for buffer in buffers:
+            PyBuffer_Release(&buffer)
 
 
 __version__ = '0.2.0'